# apps/api/stratify-data-api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stratify Data API",
    "description": "A REST API for accessing stock market data powered by Yahoo Finance and Stooq.",
    "version": "0.4.0"
  },
  "paths": {
...
    "/market/top-gainers": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Get Top Gainers",
        "operationId": "get_top_gainers_market_top_gainers_get",
        "parameters": [
          {
            "name": "minimumPercentageChange",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 3.0,
              "title": "Minimumpercentagechange"
            }
          },
          {
            "name": "minimumVolume",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1000000,
              "title": "Minimumvolume"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteListGetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/market/top-losers": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Get Top Losers",
        "operationId": "get_top_losers_market_top_losers_get",
        "parameters": [
          {
            "name": "minimumPercentageChange",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": -3.0,
              "title": "Minimumpercentagechange"
            }
          },
          {
            "name": "minimumVolume",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1000000,
              "title": "Minimumvolume"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteListGetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/market/most-active": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Get Most Active",
        "operationId": "get_most_active_market_most_active_get",
        "parameters": [
          {
            "name": "minimumMarketCap",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 2000000000,
              "title": "Minimummarketcap"
            }
          },
          {
            "name": "minimumVolume",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1000000,
              "title": "Minimumvolume"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteListGetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OHLCV": {
        "properties": {
          "open": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open"
          },
          "high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "High"
          },
          "low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low"
          },
          "close": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close"
          },
          "volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "change": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change"
          },
          "changePercent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changepercent"
          }
        },
        "type": "object",
        "title": "OHLCV"
      },
      "PriceDetails": {
        "properties": {
          "currentPrice": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentprice"
          },
          "dayTradingActivity": {
            "$ref": "#/components/schemas/OHLCV"
          }
        },
        "type": "object",
        "required": [
          "dayTradingActivity"
        ],
        "title": "PriceDetails"
      },
      "QuoteItem": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "marketState": {
            "type": "string",
            "title": "Marketstate"
          },
          "assetType": {
            "type": "string",
            "title": "Assettype"
          },
          "priceDetails": {
            "$ref": "#/components/schemas/PriceDetails"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "marketState",
          "assetType",
          "priceDetails"
        ],
        "title": "QuoteItem"
      },
      "QuoteListGetResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/QuoteItem"
            },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "QuoteListGetResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}